Fixes on sensor fusion and processor of old radar events
