+++ mobility_5g_rest_api/management/commands/process_old_radar_events.py
@@ -1,0 +1,27 @@
+from datetime import datetime, timedelta
+from django.core.management.base import BaseCommand
+from mobility_5g_rest_api.models import Event, RadarEvent
+
+
+class Command(BaseCommand):
+    help = 'Processor of old radar events'
+
+    def handle(self, *args, **options):
+        print("Running Old Radar Event")
+        for radar_id in [5, 7]:
+            if radar_id == 5:
+                location = 'PT'
+            elif radar_id == 7:
+                location = 'RA'
+
+            old_radar_events = RadarEvent.objects.filter(timestamp__lt=(datetime.now() - timedelta(seconds=30)),
+                                                         radar_id=radar_id)
+            for event in old_radar_events:
+                if event.velocity >= 5:
+                    Event.objects.create(location=location,
+                                         event_type="RT",
+                                         event_class="CA",
+                                         timestamp=event.timestamp,
+                                         velocity=event.velocity)
+                event.delete()
+
